refactor: generalize fetch api and base config
- #75 TG-238
