+++ backend/src/canvasAPI/courseRequests.ts
@@ -1,0 +1,3 @@
+/**
+ * Collection of course-related queries for the Canvas API.
+ */
